[guiinfo] Implement gui controls guiinfo provider.
